[refactor] 북마크 리포지토리 구현체 impl로 이동 (#46)
