+++ src/main/java/org/recordy/server/record_stat/repository/impl/BookmarkJpaRepository.java
@@ -1,4 +1,4 @@
-package org.recordy.server.record_stat.repository;
+package org.recordy.server.record_stat.repository.impl;
 
 import org.recordy.server.record_stat.domain.BookmarkEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
